// interfaces/retailPrice.ts
export interface RetailPrice {
  ExaminDate: string;
  ExaminGradeCode: string;
  ExaminGradeName: string;
  ExaminItemCode: string;
  ExaminItemName: string;
  ExaminSpeciesCode: string;
  ExaminSpeciesName: string;
  ExaminUnit: string;
  ExaminUnitName: string;
  MaxPrice: number;
  MinPrice: number;
}